+++ packages/studio/src/components/editor/domEditOverlayGeometry.ts
@@ -37,82 +37,6 @@ export function isElementVisibleForOverlay(el: HTMLElement): boolean {
 // shapes (rectangular cards, text, full-bleed media) don't have interior holes, so this
 // doesn't bite. If ring/cutout shapes become editable targets, sample more densely or
 // hit-test against the element's actual painted geometry instead of its bounding box.
-const OCCLUSION_SAMPLE_POINTS: ReadonlyArray<readonly [number, number]> = [
-  [0.5, 0.5],
-  [0.2, 0.2],
-  [0.8, 0.2],
-  [0.2, 0.8],
-  [0.8, 0.8],
-];
-
-/**
- * Cumulative opacity of an element through its ancestors (0 if any link is ~0).
- *
- * The caller (`isElementVisibleInPreview`) treats a result `> 0.01` as an OPAQUE
- * cover that blocks the element behind it. The `0.01` cutoff (rather than a literal
- * `> 0`) ignores sub-perceptible residue — a scene parked at `opacity: 0` can compute
- * to a tiny non-zero value (rounding, an in-flight tween a frame from settling), and
- * we don't want that to register as a real cover.
- *
- * RUNTIME INVARIANT this relies on: in the current composition model scenes only ever
- * fade IN (0 → 1) and then stay; they never fade OUT and are never parked at a partial
- * mid-opacity. So any element above this threshold is genuinely an opaque, fully-painted
- * cover. A fade-OUT or a crossfade that lingers at, say, opacity 0.4 would break this:
- * `effectiveOpacity` would report `> 0.01` for a half-transparent layer and the occlusion
- * test would wrongly treat the element behind it as hidden. If the model ever grows
- * fade-out / crossfade scenes, the occlusion test needs a real coverage model (alpha
- * accumulation), not this binary cutoff.
- */
-function effectiveOpacity(el: Element | null, win: Window): number {
-  let opacity = 1;
-  let current: Element | null = el;
-  while (current) {
-    const op = Number.parseFloat(win.getComputedStyle(current).opacity);
-    if (Number.isFinite(op)) opacity *= op;
-    if (opacity <= 0.01) return 0;
-    current = current.parentElement;
-  }
-  return opacity;
-}
-
-/**
- * True when the element is actually painted on screen — what the viewer sees in
- * the preview. Extends `isElementVisibleForOverlay` (display/visibility/opacity)
- * with an OCCLUSION test: this composition stacks scenes by z-index and fades them
- * IN (never out), so an earlier scene's element stays opacity-1 yet is covered by a
- * later opaque scene.
- *
- * Walks the painted stack (`elementsFromPoint`, top→bottom) at several sample points.
- * A point "sees" the element if the element (or its subtree/ancestor) is reached
- * before any unrelated element that's effectively opaque. Transparent covers (a
- * faded-in scene still at opacity ~0) are skipped — they hit-test but don't paint.
- * If every sampled point is blocked by an opaque cover, the element is hidden.
- */
-export function isElementVisibleInPreview(el: HTMLElement): boolean {
-  if (!isElementVisibleForOverlay(el)) return false;
-  const doc = el.ownerDocument;
-  const win = doc.defaultView;
-  if (!win || typeof doc.elementsFromPoint !== "function") return true;
-  const rect = el.getBoundingClientRect();
-  if (rect.width <= 0 || rect.height <= 0) return false;
-
-  let sampledInViewport = false;
-  for (const [fx, fy] of OCCLUSION_SAMPLE_POINTS) {
-    const x = rect.left + rect.width * fx;
-    const y = rect.top + rect.height * fy;
-    if (x < 0 || y < 0 || x > win.innerWidth || y > win.innerHeight) continue;
-    sampledInViewport = true;
-    for (const hit of doc.elementsFromPoint(x, y)) {
-      if (hit === el || el.contains(hit) || hit.contains(el)) return true; // reached, uncovered
-      if (effectiveOpacity(hit, win) > 0.01) break; // opaque cover above → this point blocked
-      // transparent cover (e.g. a scene at opacity ~0) → ignore, keep descending
-    }
-  }
-  // Every in-viewport sample was blocked by an opaque cover → occluded. If nothing
-  // was testable (off-viewport), don't hide on this basis.
-  return !sampledInViewport;
-}
-
 function readPositiveDimension(value: string | null): number | null {
   if (!value) return null;
   const parsed = Number.parseFloat(value);
