+++ packages/studio/src/hooks/useGestureCommit.ts
@@ -3,6 +3,7 @@
  * Extracted from App.tsx to keep file sizes under the 600-line limit.
  */
 import { useState, useCallback, useRef, useEffect } from "react";
+import { editLog } from "../utils/editDebugLog";
 import { useGestureRecording } from "./useGestureRecording";
 import { simplifyGestureSamples } from "../utils/rdpSimplify";
 import { usePlayerStore } from "../player";
@@ -270,6 +271,9 @@ export function useGestureCommit({
   }, [gestureRecording, showToast, isGestureRecordingRef, domEditSessionRef]);
 
   const handleToggleRecording = useCallback(() => {
+    editLog("gesture", gestureStateRef.current === "recording" ? "stop" : "start", {
+      id: domEditSessionRef.current.domEditSelection?.id,
+    });
     if (gestureStateRef.current === "recording") {
       void stopAndCommitRecording();
       return;
